Replace UART base addr variables with macros
Storing the base addresses for the UARTs in global variables is not
working correctly. I am not completely sure what the issue is but even
after calling sdTerminalInit() - potato_uart_base and qemu_uart_base
sometimes are 0. This makes pForth hang at {qemu,potato}UARTTXFull() as
it reads from the wrong location. Work around this by using constants
for the base address. This is not really a problem as the addresses do
not change anyway.

Signed-off-by: Jordan Niethe <jniethe5@gmail.com>