Check TX has room before writing
When the base addresses were stored in global variables, the weirdness
around the values sometimes being 0 made me take out the TX space
checks. This was as while the TX check would hang pForth, after
transmitting for a while, the base address would get set correctly and
we would see output. This worked for simulation where everything is very
slow, but did not work on the actually FPGA. Now that the issue with the
base addresses is fixed, add the checks in.

Signed-off-by: Jordan Niethe <jniethe5@gmail.com>